Module cleaning: rename network.py to system.py, delete probability_law.py
- Rename src/gems/study/network.py → system.py (preserving git history)
- Update all import references from gems.study.network → gems.study.system
- Rename local variable `network` → `system` in folder.py and tests
- Update comments/variable names in optim_config/parsing.py
- Delete src/gems/model/probability_law.py (unused stub)

Closes #74
