+++ tests/e2e/functional/test_libs_yaml_system_yaml.py
@@ -45,7 +45,7 @@ from gems.model.parsing import InputLibrary, parse_yaml_library
 from gems.model.resolve_library import resolve_library
 from gems.simulation import BlockBorderManagement, TimeBlock, build_problem
 from gems.study.data import DataBase
-from gems.study.network import System
+from gems.study.system import System
 from gems.study.parsing import InputSystem, parse_yaml_components
 from gems.study.resolve_components import (
     build_data_base,
