+++ tests/e2e/functional/test_study_from_folder.py
@@ -7,9 +7,9 @@ from gems.study.folder import load_study, run_study
 def test_load_study():
     study_dir = Path(__file__).parent / "studies" / "7_4"
 
-    network, database = load_study(study_dir)
-    assert len(network.components) == 12
-    assert len(network.connections) == 11
+    system, database = load_study(study_dir)
+    assert len(system.components) == 12
+    assert len(system.connections) == 11
     assert len(database._data) == 76
